streamlit-absensi/requirements.txt
streamlit==1.29.0
paho-mqtt==1.6.1